Refactor map-glyphs into two functions to allow better control over characters.
The CALL-WITH-GLYPH-INFO function works by accepting a starting character
and expecting a next character from the called function. It will do the
layout bookkeeping, but leave the iteration order or character generation
up to the user.

Note that this is a part of what a proper layouting engine would be doing,
just restricted to only dealing with LTR text. I think it would be better
to leave almost all of this stuff out of a glyph rendering system, and
instead rely on an external system to do all the layouting decisions.